possible fixes for 10.1.7.51972
I have removed and commented on all the errors that occur when starting the UI so far. However, since I am currently using the free version and can only level up to level 20, I cannot say for sure if everything is working perfectly. I'm not sure yet if I will come back and completely fix the UI, but for now, it should be up and running.